requirements.txt
pandas
scikit-learn
scipy
numpy
alibi
tldextract
pyarrow
click
torch
setuptools
redis
pytest
feature-engineering-polars